Fix windows build by moving references to SIGUSR2 into a file not built on windows.
