+++ command/commands_windows.go
@@ -1,0 +1,8 @@
+// +build windows
+
+package command
+
+// MakeSigUSR2Ch does nothing useful on Windows.
+func MakeSigUSR2Ch() chan struct{} {
+	return make(chan struct{})
+}
